TakasumiBOTworknotification-main/requirements.txt
discord.py
python-dotenv
requests
asyncio
psutil